requirements.txt
blinker==1.8.2
cachelib==0.13.0
certifi==2024.6.2
click==8.1.7
DateTime==5.5
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Flask-Session==0.8.0
Flask-Uploads==0.2.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgspec==0.18.6
packaging==24.1
PyJWT==2.8.0
pymongo==4.7.3
python-dotenv==1.0.1
pytz==2024.1
uuid==1.30
Werkzeug==3.0.3
zope.interface==6.4.post2
openai==0.28.0
